+++ luirary/src/Token.java
@@ -2,12 +2,12 @@ public class Token
 {
     public enum Type {OPERAND, OPERATOR}
     private Type type;
-    private int operand;
+    private double operand;
     private char operator;
     private int precedence = -1;
     private boolean rightAssociativity = false;
 
-    public Token(int operand)
+    public Token(double operand)
     {
         type = Type.OPERAND;
         this.operand = operand;
@@ -42,7 +42,7 @@ public class Token
         }
     }
 
-    public void setOperand(int operand)
+    public void setOperand(double operand)
     {
         this.operand = operand;
     }
@@ -57,7 +57,7 @@ public class Token
         return this.type;
     }
 
-    public int getOperand()
+    public double getOperand()
     {
         return this.operand;
     }
